Adds initialization of map on room-page
